auth service: Add user to loginSuccess mutation
