+++ apps/obs-wcgop-optecs/src/_store/index.ts
@@ -1,16 +1,23 @@
 import Vue from 'vue';
-import Vuex from 'vuex';
+import Vuex, { StoreOptions } from 'vuex';
 
-import { alert } from './alert.module';
-import { auth } from './auth.module';
+import { alert } from '@/_store/alert.module';
+import { auth } from '@/_store/auth.module';
+
+import { RootState } from '@/_store/types/types';
+
+// Primary app store: following examples at https://codeburst.io/vuex-and-typescript-3427ba78cfa8
 
 Vue.use(Vuex);
 
-export const store = new Vuex.Store({
+const store: StoreOptions<RootState> = {
+  state: {
+    version: '1.0.0'
+  },
   modules: {
     alert,
     auth
   }
-});
+};
 
-export default store;
+export default new Vuex.Store<RootState>(store);
