+++ apps/obs-wcgop-optecs/src/main.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import App from './App.vue';
 import router from './router';
-import store from './store';
+import store from './_store';
 import './registerServiceWorker';
 
 // The quasar stylus imports are somewhat broken, so import css here
